[UPDATE]: organization of code into modules
